README.md
If you're looking to join one of the fastest-growing cryptocurrency exchanges with incredible trading rewards, then using the Bitget referral code is a must. With the Bitget referral code BGB101, new users from the USA, Europe, and many other countries can enjoy a 50% discount on trading fees, unlock up to 6200 USDT in bonuses, and benefit from exclusive signup promotions in 2025.